cart features, disabled auth

+++ src/main/java/com/extron/MyFirstECom/Repository/UserRepo.java
@@ -3,6 +3,7 @@ package com.extron.MyFirstECom.Repository;
 import com.extron.MyFirstECom.Model.Users;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepo extends JpaRepository<Users, Integer> {
+
+public interface UserRepo extends JpaRepository<Users, Long> {
     Users findByUsername(String username);
 }
